Added a faster stifness matrix assemby routine.
This version is approximately 3 times faster than
the previous. The time taken to assemble the stiffness
matrices, for standard fem problems, no longer dominants
the cpu time required to perform one time step.

This is be no means the final say in matrix assembly
routines, and I welcome any suggestions. There are
certainly improvements which can be made.

Within the assembly routines, timing information can be
obtained for the specific tasks. This should provide
useful information for profiling. The timing information
is activated by setting the debug level to be >= 1, i.e.

<param name="journal-level-branch.debug.StgFEM"> 1 </param>