The assembly routine is broken into three phases. 	1) build the operator (requires all element stiff. mats to be assembled) 	2) build the corrections for the rhs associated with the operator 		(requires elements with nodes which have constraints to be assembled) 	3) build the corrections for the rhs associated with the transpose of the operator 		(requires elements with nodes which have constraints to be assembled)
This approach will build certain element stiffness matrices several times. The overhead
for doing this is small. It is possible to do steps 2) and 3) within the one function,
it would require embedding more "if" statements within the element loop.
The assembly code (as it is) is clean, and easy to maintain and fast, despite a small amount
of redundant calculations.