Cleaned up the information reported to stdout.
* Modified the journaling statements in the files below.

* Most of the places where I made the modifications were either
print redundant information on every processor, or simply logging
time information. In the latter case, I just report the min and
max times.

* I don't report which cpu which produced the min/max time, but
if ever the difference between the min and max is large you
know that the function in question has an obvious sync. problem
and further examination is obviously required.