** As requested by LM **
Added more monitoring for the picard solver. We now report each of the
norms used to determine if convergence has occurred and how far each
norm is from achieving convergence. The latter quantity is report as
a percentage.

The monitor can be activated via
  --components.NAME.picard_ActivateMonitor=true
where NAME is the textual name of your component.
For underworld NAME=stokesEqn

The monitor format looks like
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
  SLE_NS  it       |F|                 |F|/|F0|             |X1-X0|/|X1|
  SLE_NS  0000     3.1275e+06 [0%]     1.0000e+00 [0%]      1.0000e+00 [0%]
  SLE_NS  0001     8.4142e+05 [1%]     2.6904e-01 [7%]      2.0442e-02 [21%]
   ...
  SLE_NS  0022     1.3987e+00 [11%]     4.4722e-07 [79%]      1.8485e-08 [97%]
  SLE_NS  0023     9.0149e-01 [12%]     2.8825e-07 [82%]      9.1613e-09 [100%]
Nonlinear solve converged due to CONVERGED_PNORM_RELATIVE
<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<