I have added a 	Bool pointSmoothing
variable to the SwarmViewer which toggles GL_POINT_SMOOTH when visualizing a swarm. This is
useful when there are not too many particles and the spaces between them become more visible
and the particle shapes are obvious.

I don't know how expensive it would be to render the large arrays of material points with this
option so I have left it off by default.

Louis