added additional function to the element type classes to evaluate the jacobian determinant across element surfaces. functions is supplied with a self ptr, the mesh, a list of nodes which define the face and the normal dimension index.
note that this function is only implemented for element types where the nodes are defined on the
element surface, as for element types where the nodes are defined as being internal to the
elements the fluxes evaluated at these nodes will be spurious. for these element types, the
jacobian surface function is passed as the default element type parent class function, which
returns an error.