Testing revealed an inconsistency in IndexSet_Merge_AND() function when first IndexSet provided was longer than the second, due to nature of bit operations used.
Have decided to resolve inconsistency by changing code so that in the case that the first list is longer than the second, all entries after the end of the second list are set to zero. I believe this is most logical for an AND operation. Have updated header file description of the function to explain this. Don't believe it'll cause an issue with other parts of the code, since I don't expect this function has been used with different length lists before.