Optimised GNx memory allocation.
The GNx array was allocated and freed every element assembly.... slow
Now it's only alloc if the GNx array needs to be increase in size, i.e. if the number of nodes per elements change.

I've moved the GNx ptr onto the StiffnessMatrixTerm abstract class, could do the same with Ni later.