modifications to correct bug encountered where swarm checkpoint files contain no particles.
unfortunately, hdf5 does not want to play with empty files.  the solution is for each created file to have an attribute which specifies how many particle the file
contains.  this is then used to skip files which contain no particles.  also, files without particles simply contain this attribute, and do not contain any datasets.