Fixed the problem with the Comm component causing the CommSuite_TestRemNbrs test to fail for 4+ procs.
Turned out it was a bug in Comm_RemoveNeighbours - this requires an array of _local_ indices of nbrs to be passed in, but then these needs to be converted to _global_ ranks when they are removed from the
global to local mapping table.

Also, added some doc comments to explain this, and made a few relevant variable names a bit more self-documenting.