committing script/createDocs.py, doc/headerfilter.py and doc/Doxyfile into a clean repo to get around merge changes.
The changes to these files are as follows:
createDocs.py:
added 'scons' as a dir to ignore in creating componentCodex

headerfilter.py:
rewrote large sections so that it correctly converts our C header files
into a C++ header file to be read by Doxygen.
It can handle finding the parent class, and setting up all the inputs.
It sets all the functions in the header file to public functions
in a struct.

Doxyfile:

Changed a few of the options.
Turned on verbatim_headers and Source_browser which allow the .h and .c
files to be visible respectively
turned off filter_headers which was causing the files to be output as
the filtered c++ file rather than the originals.
removing the filter_patterns value as this is set in the python file
createDocs and was being overwritten anyway. Leaving this in could cause confusion
changed a few more options which defined which data to display.
but that's all the important  ones for the moment